+ bugfix
git-svn-id: http://agilefant.googlecode.com/svn/trunk@2851 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/business/impl/BacklogBusinessImpl.java
@@ -117,7 +117,7 @@ public class BacklogBusinessImpl extends GenericBusinessImpl<Backlog> implements
     public void addAssignees(int backlogId, Set<Integer> userIds) {
         Backlog backlog = this.retrieve(backlogId);
         this.assignmentBusiness.addMultiple(backlog, userIds,
-                ExactEstimate.ZERO, (short) 100);
+                ExactEstimate.ZERO, 100);
     }
 
 }
